Testing transformation matrix (should cover the results of the analysis object). moved write input to the new file searated analysis and solution.py
